refactor: consolidate OpenAICompatibleProvider into _openai_common
