+++ src/pincer/llm/__init__.py
@@ -1,3 +1,4 @@
+from pincer.llm._openai_common import OpenAICompatibleProvider
 from pincer.llm.base import (
     BaseLLMProvider,
     ImageContent,
@@ -8,7 +9,6 @@ from pincer.llm.base import (
     ToolResult,
 )
 from pincer.llm.ollama_provider import OllamaProvider
-from pincer.llm.openai_compatible_provider import OpenAICompatibleProvider
 
 __all__ = [
     "BaseLLMProvider",
